requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
requests==2.31.0
python-imagesearch==1.3.0
pillow==10.2.0
pytesseract==0.3.10
setuptools==69.0.3
PyQt5==5.15.10
